% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_ccre_enrich.R
\name{analyze_ccre_enrich}
\alias{analyze_ccre_enrich}
\title{Analyze the enrichment of cCREs.}
\usage{
analyze_ccre_enrich(snps_loc, assembly = "hg38", show_p = FALSE)
}
\arguments{
\item{snps_loc}{Required. The genomic locations of SNPs, the format should be like: "chr1:1014863".}

\item{assembly}{Optional. "hg19" and hg38" can be selected. Default is "hg38".}

\item{show_p}{Optional. Show p-values or "***" (significant) on output plot. Default is FALSE.}
}
\value{
A plot.
}
\description{
Analyze the enrichment of cCREs.
}
\examples{
analyze_ccre_enrich(snps_loc)
}
