% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_gene_go.R
\name{analyze_gene_go}
\alias{analyze_gene_go}
\title{Return the go enrichment results of genes.}
\usage{
analyze_gene_go(snp_genes, output_type = "dotplot")
}
\arguments{
\item{snp_genes}{Required. The annotated genes of input SNPs.}

\item{output_type}{Optional. Show dotplot/barplot of GO enrichment results. Default is dotplot.}
}
\value{
A plot.
}
\description{
Return the go enrichment results of genes.
}
\examples{
analyze_gene_go(snps_genes)
}
