% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_gwas_enrich.R
\name{analyze_gwas_enrich}
\alias{analyze_gwas_enrich}
\title{Return the enrichment of GWAS studies.}
\usage{
analyze_gwas_enrich(snps_loc, assembly = "hg38", show_p = FALSE)
}
\arguments{
\item{snps_loc}{Required. The locations of SNPs, the format should be "chr1:1014863".}

\item{assembly}{Optional. "hg19" and hg38" can be selected. Default is "hg38".}

\item{show_p}{Optional. Show p-values or "***" (significant) on output plot. Default is FALSE.}
}
\value{
A plot.
}
\description{
Return the enrichment of GWAS studies.
}
\examples{
analyze_gwas_enrich(snps_loc)
}
