% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_batch_SNP_info.R
\name{get_batch_SNP_info}
\alias{get_batch_SNP_info}
\title{Return information of a batch of SNPs (Annotated by query_snps())}
\usage{
get_batch_SNP_info(snps, input_type = "rsID", eqtl_tissue = "Whole_Blood")
}
\arguments{
\item{snps}{Required.}

\item{input_type}{Optional. The assembly version of the input SNP. "rsID", "hg19" and "hg38" can be selected. Default is "rsID".}

\item{eqtl_tissue}{Optional. Tissue ID of the tissue of interest. Default is "Whole_Blood".}
}
\value{
A data.frame.
}
\description{
Return information of a batch of SNPs (Annotated by query_snps())
}
\examples{
batch_info_table <- get_batch_SNP_info(snps = c("rs1891906" "rs10"), input_type="rsID", eqtl_tissue="Whole_Blood")
}
