% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_batch_vep.R
\name{get_batch_vep}
\alias{get_batch_vep}
\title{Return annotation by VEP}
\usage{
get_batch_vep(snps, input_type = "rsID")
}
\arguments{
\item{snps}{Required.}

\item{input_type}{Optional. The type of the input SNP. "rsID", hg19" or "hg38" can be selected. Default is "rsID".}
}
\value{
A data.frame of VEP annotation
}
\description{
Return annotation by VEP
}
\examples{
vep_anno <- get_batch_vep(c("rs56116432", "rs10040658"), input_type = "rsID") # List of annotations
}
