% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genes_loc.R
\name{get_genes_loc}
\alias{get_genes_loc}
\title{Return the genomic locations of gene symbols}
\usage{
get_genes_loc(genes, assembly = "hg19")
}
\arguments{
\item{genes}{Required. Gene Symbol(s), eg. "GP1BB"}

\item{assembly}{Optional. Output UCSC Assembly versions. "hg19" or "hg38" can be selected. Default is "hg19".}
}
\value{
A data.frame including the genomic location.
}
\description{
Return the genomic locations of gene symbols
}
\examples{
get_genes_loc(c('GP1BB', 'TBX1'))

}
