% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_LD.R
\name{get_LD}
\alias{get_LD}
\title{Title get_LD: Get the genomic location of one gene symbol
It should be noted that LDlinkR is called here, so a Personal Access Token must be required. Please refer to https://cran.r-project.org/web/packages/LDlinkR/vignettes/LDlinkR.html for details.}
\usage{
get_LD(snps, pop = "CEU", r2d = "r2", assembly = "hg19", plot_heatmap = TRUE)
}
\arguments{
\item{snps}{Required. between 1 - 10 variants, using an rsID or chromosome coordinate.}

\item{pop}{Optional. Default = “CEU”. list_pop() return all the options.}

\item{r2d}{Optional. Default = “r2”. Either “r2” for LD R2 (R-squared) or “d” for LD D’ can be selected.}

\item{assembly}{Optional. Output UCSC Assembly versions. "hg19" or "hg38" can be selected. Default is "hg19".}

\item{plot_heatmap}{Optional. Whether to output the heatmap of the LD block. Default is TRUE.}
}
\value{
A data.frame of pairwise linkage disequilibrium statistics. Desired output can be based on estimates of R2 or D’.
}
\description{
Title get_LD: Get the genomic location of one gene symbol
It should be noted that LDlinkR is called here, so a Personal Access Token must be required. Please refer to https://cran.r-project.org/web/packages/LDlinkR/vignettes/LDlinkR.html for details.
}
\examples{
 ld_block <- get_LD(c("chr7:24962419", "rs114", "rs127", "rs7805287", "rs60676332", "rs10239961"), assembly = "hg19")

}
