% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loop_gene.R
\name{get_loop_gene}
\alias{get_loop_gene}
\title{Return 3D-interacting genes of a SNP}
\usage{
get_loop_gene(snp, input_type = "rsID", output_type = "core")
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The type of the input SNP. "rsID", hg19" or "hg38" can be selected. Default is "rsID".}

\item{output_type}{Optional. The type of the output matrix. "core" or "full" can be selected. The output of "core" is subset of "full".Default is "core".}
}
\value{
A data.frame of 3D-interacting genes.
}
\description{
Return 3D-interacting genes of a SNP
}
\examples{
loop_genes <- get_loop_gene("rs1059196")
}
