% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_ccre.R
\name{get_snp_ccre}
\alias{get_snp_ccre}
\title{Return the cCREs (candidate cis-regulatory elements) that the SNPs located in.}
\usage{
get_snp_ccre(snp, input_type = "rsID", output_assembly = "hg19")
}
\arguments{
\item{snp}{Required. The chromosome coordinate or rsID of a SNP, eg. "chr22:19712094"}

\item{input_type}{Optional. The type of the input SNP. "rsID", hg19" or "hg38" can be selected. Default is "rsID".}

\item{output_assembly}{Optional. The UCSC Assembly versions of chromosome coordinates of cCREs. "hg19" or "hg38" can be selected. Default is "hg19".}
}
\value{
A data.frame of overlapped cCREs.
}
\description{
Return the cCREs (candidate cis-regulatory elements) that the SNPs located in.
}
\examples{
cCRE_intersect <- get_snp_ccre(snp="rs1059196")
}
