% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_eqtl.R
\name{get_snp_eqtl}
\alias{get_snp_eqtl}
\title{Return GTEx significant single tissue eQTLs.}
\usage{
get_snp_eqtl(snp, input_type = "rsID", eqtl_tissue = "Whole_Blood")
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The type of the input SNP. "rsID", "hg19" or "hg38" can be selected. Default is "rsID".}

\item{eqtl_tissue}{Optional. Tissue ID of the tissue of interest. Default is "Whole_Blood".}
}
\value{
A data.frame of associated phenotypes and reserach of the SNP.
}
\description{
Using the \href{https://gtexportal.org/api/v2/association/singleTissueEqtl}{GTEx api} to generate the information.
}
\examples{
gtex_data <- get_snp_eqtl(snp="rs10040658", eqtl_tissue = "Whole_Blood")
}
