% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_loc.R
\name{get_snp_loc}
\alias{get_snp_loc}
\title{Convert SNP rsID to genomic locations.}
\usage{
get_snp_loc(snp, assembly = "hg19")
}
\arguments{
\item{snp}{Required.}

\item{assembly}{Optional. The assembly version of the input SNP. hg19" and "hg38" can be selected. Default is "hg19".}
}
\value{
A character.
}
\description{
Convert SNP rsID to genomic locations.
}
\examples{
get_snp_loc("rs1059196")
}
