% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_motif.R
\name{get_snp_motif}
\alias{get_snp_motif}
\title{Return SNP-disrupted motifs}
\usage{
get_snp_motif(snp, input_type = "rsID")
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The type of the input SNP. "rsID", hg19" or "hg38" can be selected. Default is "rsID".}
}
\value{
A data.frame of disrupted motifs.
}
\description{
Return SNP-disrupted motifs
}
\examples{
disrupted_motifs <- get_snp_motif("chr1:109676139", "hg38")
}
