% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_rsID.R
\name{get_snp_rsID}
\alias{get_snp_rsID}
\title{Convert the chromosome coordinate of a SNP to rsID}
\usage{
get_snp_rsID(snp_loc, assembly = "hg19")
}
\arguments{
\item{snp_loc}{Required. The chromosome coordinate of a SNP, eg. "chr22:19712094"}

\item{assembly}{Optional. UCSC Assembly versions. "hg19" or "hg38" can be selected. Default is "hg19".}
}
\value{
The rsID of the SNP.
}
\description{
Convert the chromosome coordinate of a SNP to rsID
}
\examples{
get_snp_rsID("chr22:19712094") # hg19

}
