% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg19tohg38.R
\name{hg19tohg38}
\alias{hg19tohg38}
\title{Convert hg38 to hg19 coordinate.}
\usage{
hg19tohg38(snp_loc)
}
\arguments{
\item{snp_loc}{The hg38 chromosome coordinate.}
}
\value{
A character. eg. "chr22:19724571".
}
\description{
Convert hg38 to hg19 coordinate.
}
\examples{
hg38tohg19('chr22:19724571')
}
