% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hg38tohg19.R
\name{hg38tohg19}
\alias{hg38tohg19}
\title{Convert hg38 to hg19 coordinate.}
\usage{
hg38tohg19(snp_loc)
}
\arguments{
\item{snp_loc}{The hg38 chromosome coordinate.}
}
\value{
A character. eg. "chr22:19724571".
}
\description{
Convert hg38 to hg19 coordinate.
}
\examples{
hg38tohg19('chr22:19724571')
}
