% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_affect_gene.R
\name{plot_affect_gene}
\alias{plot_affect_gene}
\title{Return barplots of affected genes of SNPs.}
\usage{
plot_affect_gene(data, plot_type = "merged", show_num = 7)
}
\arguments{
\item{data}{Required. The annotation results from VEP.}

\item{plot_type}{Optional."gene", "snp", "all" and "merged" can be selected. Default is "rsID".}

\item{show_num}{Optional. The number of affected genes shown on the plot. Default is 7.}
}
\value{
A data.frame.
}
\description{
Return barplots of affected genes of SNPs.
}
\examples{
pdf("affected_genes.pdf", height = 7, width = 8)
plot_affect_gene(data, plot_type = "merged")
dev.off()
}
