% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_batch_AF.R
\name{plot_batch_AF}
\alias{plot_batch_AF}
\title{Return boxplots of allele frequency.}
\usage{
plot_batch_AF(data, version = "1000G")
}
\arguments{
\item{data}{Required. The annotation results from VEP.}

\item{version}{Optional."1000G" or "gnomAD" can be selected;. Default is "1000G".}
}
\value{
A boxplot.
}
\description{
Return boxplots of allele frequency.
}
\examples{
plot_AF(data, version = "1000G")
}
