% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_batch_titv.R
\name{plot_batch_titv}
\alias{plot_batch_titv}
\title{Return a barplot of ti/tv ratio.}
\usage{
plot_batch_titv(data)
}
\arguments{
\item{data}{Required. The annotation results from VEP (Upload only).}
}
\value{
A barplot.
}
\description{
Return a barplot of ti/tv ratio.
}
\examples{
plot_batch_titv(data)
}
