% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loop_gene.R
\name{plot_loop_gene}
\alias{plot_loop_gene}
\title{Return 3D-loop genes.}
\usage{
plot_loop_gene(
  snp,
  input_type = "rsID",
  output_assembly = "hg19",
  show_cells = 3
)
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The assembly version of the input SNP. "rsID", "hg19" and "hg38" can be selected. Default is "rsID".}

\item{output_assembly}{Optional. The UCSC Assembly versions of chromosome coordinates of SNPs. "hg19" or "hg38" can be selected. Default is "hg19".}

\item{show_cells}{Optional. The number of cell types shown in the figure. Default is 3 (Top 3 cell types). The names of cell types are available as well.}
}
\value{
A plot.
}
\description{
Return 3D-loop genes.
}
\examples{
plot_loop_snps(snp="rs10")
plot_loop_snps(snp="rs10", output_assembly='hg38', show_cells=c("VentricleLeft", "Caki2", "HepG2"))

}
