% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_loop_snp.R
\name{plot_loop_snp}
\alias{plot_loop_snp}
\title{Plot 3D-interating SNPs.}
\usage{
plot_loop_snp(
  snp,
  input_type = "rsID",
  pop,
  output_assembly = "hg19",
  show_cells = 3,
  curvature = 0.8
)
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The assembly version of the input SNP. "rsID", "hg19" and "hg38" can be selected. Default is "rsID".}

\item{pop}{Required. Population.}

\item{output_assembly}{Optional. The UCSC Assembly versions of chromosome coordinates of outputs. "hg19" or "hg38" can be selected. Default is "hg19".}

\item{show_cells}{Optional. The number of cell types shown in the figure. Default is 3 (Top 3 cell types). The names of cell types are available as well.}

\item{curvature}{Optional. The "curvature" parameter of `geom_curve`. Default is "0.8".}
}
\value{
A plot.
}
\description{
Plot 3D-interating SNPs.
}
\examples{
plot_loop_snp(snp="rs10", pop="AFR")
plot_loop_snp(snp="rs10", pop="AFR", output_assembly='hg38')
plot_loop_snp(snp="rs10", pop = pop, show_cells = c("Ventricle_Right", "Spleen"))
plot_loop_snp(snp='rs10040658', pop = "AFR", show_cells = c("GM12878", "Lung"))
}
