% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_physcores.R
\name{plot_physcores}
\alias{plot_physcores}
\title{Return a table and a line chart of the physcores of +/-10bp of SNPs}
\usage{
plot_physcores(snps, colors = "default", input_type = "rsID")
}
\arguments{
\item{snps}{Required.}

\item{colors}{Optional. Colors of lines. If not specified, default is to use the colors of `rainbow()`.}

\item{input_type}{Optional. The assembly version of the input SNP. "rsID", "hg19" and "hg38" can be selected. Default is "rsID".}
}
\value{
A data.frame of physcores and a line chart.
}
\description{
Return a table and a line chart of the physcores of +/-10bp of SNPs
}
\examples{
snps <- c("rs2769466", "rs4661232")
physcores_table <- plot_physcores(snps, colors = c("blue", "orange"))
}
