% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snp_circos.R
\name{plot_snp_circos}
\alias{plot_snp_circos}
\title{Return the circos plot.}
\usage{
plot_snp_circos(
  snp_info_table,
  output_assembly = "hg19",
  window_size = 2e+05,
  savefile = "circos.pdf"
)
}
\arguments{
\item{snp_info_table}{Required. The output data.frame of `query_snp()`.}

\item{output_assembly}{Optional. The version of chromosome coordinates of output circos plots. Default is "hg19".}

\item{window_size}{Optional. Window size around the SNP displayed. Default is 2e5.}

\item{savefile}{Optional. The filename of the circos plot.}
}
\value{
A circos plot.
}
\description{
Return the circos plot.
}
\examples{
info_table <- query_snp("rs1891906")
plot_snp_circos(snp_info_table=info_table, window_size = 1e5, savefile="~/circos.pdf")
}
