% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snp_motif.R
\name{plot_snp_motif}
\alias{plot_snp_motif}
\title{Return the disrupted motif sequences.}
\usage{
plot_snp_motif(snp, input_type = "rsID", tf = "all")
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The assembly version of the input SNP. "rsID", "hg19" and "hg38" can be selected. Default is "rsID".}
}
\value{
A plot.
}
\description{
Return the disrupted motif sequences.
}
\examples{
plot_snp_motif(snp="rs10040658")
}
