% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_snp.R
\name{query_snp}
\alias{query_snp}
\title{Return information of a SNP}
\usage{
query_snp(snp, input_type = "rsID", eqtl_tissue = "Whole_Blood")
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The assembly version of the input SNP. "rsID", "hg19" and "hg38" can be selected. Default is "rsID".}

\item{eqtl_tissue}{Optional. Tissue ID of the tissue of interest. Default is "Whole_Blood".}
}
\value{
A data.frame.
}
\description{
Return information of a SNP
}
\examples{
info_table <- query_snp("chr1:1014863 ", input_type="hg38")
}
