% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screenshot_GenomeBroswer.R
\name{screenshot_GenomeBroswer}
\alias{screenshot_GenomeBroswer}
\title{Return the Genome Browser screenshot of a SNP.}
\usage{
screenshot_GenomeBroswer(
  snp,
  input_type = "rsID",
  filename = "screenshot.pdf",
  window_size = 20
)
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The type of the input SNP. "rsID", hg19" or "hg38" can be selected. Default is "rsID".}

\item{filename}{Optional. The filename of the output screenshot.}

\item{window_size}{Optional. The window size around the SNP.}
}
\value{
A PDF screenshot.
}
\description{
Return the Genome Browser screenshot of a SNP.
}
\examples{
screenshot_GenomeBroswer("chr22:19712094", input_type = "hg19", "~/test.pdf")

}
