% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_vep.R
\name{summary_vep}
\alias{summary_vep}
\title{Return barplots of affected genes of SNPs.}
\usage{
summary_vep(vep_file, output = "summmary_vep.txt")
}
\arguments{
\item{data}{Required. The annotation results from VEP.}

\item{plot_type}{Optional."gene", "snp", "all" and "merged" can be selected. Default is "rsID".}

\item{show_num}{Optional. The number of affected genes shown on the plot. Default is 7.}
}
\value{
A data.frame.
}
\description{
Return barplots of affected genes of SNPs.
}
