% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_loop_snp.R
\name{get_loop_snp}
\alias{get_loop_snp}
\title{Return 3D-interacting SNPs of a SNP}
\usage{
get_loop_snp(snp, input_type = "rsID")
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The type of the input SNP. "rsID", hg19" or "hg38" can be selected. Default is "rsID".}
}
\value{
A data.frame of 3D-interacting SNPs
}
\description{
Return 3D-interacting SNPs of a SNP
}
\examples{
loop_snps <- get_loop_snp("rs1059196")
}
