% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_gwas.R
\name{get_snp_gwas}
\alias{get_snp_gwas}
\title{Return the associated phenotypes and related research (from GWAS catalog) of the SNP.}
\usage{
get_snp_gwas(snp, input_type = "rsID", output_type = "core")
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The type of the input SNP. "rsID", hg19" or "hg38" can be selected. Default is "rsID".}

\item{output_type}{Optional. The type of the output matrix. "core" or "full" can be selected. The output of "core" is subset of "full".Default is "core".}
}
\value{
A data.frame of associated GWAS phenpotypes.
}
\description{
Return the associated phenotypes and related research (from GWAS catalog) of the SNP.
}
\examples{
gwas_info <- get_snp_gwas("rs1891906")
}
