% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_gwas.R
\name{get_snp_gwas}
\alias{get_snp_gwas}
\title{Title get_snp_eqtl: Return the associated phenotypes and related research (from GWAS catelog) of the SNP.}
\usage{
get_snp_gwas(snp, input_type = "rsID")
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The type of the input SNP. "rsID", hg19" or "hg38" can be selected. Default is "rsID".}
}
\value{
A data.frame of associated GWAS phenpotypes.
}
\description{
Title get_snp_eqtl: Return the associated phenotypes and related research (from GWAS catelog) of the SNP.
}
\examples{
gwas_info <- get_snp_gwas("rs1891906")
}
