% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_batch_cCREs.R
\name{plot_batch_cCREs}
\alias{plot_batch_cCREs}
\title{Return barplots of cCREs that SNPs overlapped.}
\usage{
plot_batch_cCREs(data, assembly = "hg38", show_unclassified = FALSE)
}
\arguments{
\item{data}{Required. The annotation results from VEP.}

\item{assembly}{Optional. "hg19" and hg38" can be selected. Default is "hg38".}

\item{show_unclassified}{Optional. Whether to display SNPs without overlapping cCREs.}
}
\value{
A data.frame.
}
\description{
Return barplots of cCREs that SNPs overlapped.
}
\examples{
plot_batch_cCREs(data, assembly = "hg38", show_unclassified=FALSE)
}
