% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_batch_consequence.R
\name{plot_batch_consequence}
\alias{plot_batch_consequence}
\title{Return barplots of consequences.}
\usage{
plot_batch_consequence(data, show_num = 7, colors = "default")
}
\arguments{
\item{data}{Required. The annotation results from VEP.}

\item{show_num}{Optional. The number of affected genes shown on the plot. Default is 7.}

\item{colors}{Optional. Default is `ggsci::pal_igv(alpha = 0.8)(show_num)`.}
}
\value{
Barplots.
}
\description{
Return barplots of consequences.
}
\examples{
plot_consequence(data, show_num = 10, colors = rainbow(10))
}
