% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_batch_gwas.R
\name{plot_batch_gwas}
\alias{plot_batch_gwas}
\title{Return barplots of associated GWAS phenotypes of SNPs.}
\usage{
plot_batch_gwas(data, assembly = "hg38", show_num = 5)
}
\arguments{
\item{data}{Required. The annotation results from VEP.}

\item{assembly}{Optional. The assembly version of the input SNPs. "hg19" and "hg38" can be selected. Default is "hg38". Default is "hg38".}

\item{show_num}{Optional. The number of associated phenotypes shown on the plot. Default is 5.}
}
\value{
A barplot.
}
\description{
Return barplots of associated GWAS phenotypes of SNPs.
}
\examples{
plot_batch_gwas(data)
}
