% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_snp_network.R
\name{plot_snp_network}
\alias{plot_snp_network}
\title{Return the network.}
\usage{
plot_snp_network(snp, input_type = "rsID", colors = "default")
}
\arguments{
\item{snp}{Required.}

\item{input_type}{Optional. The assembly version of the input SNP. "rsID", "hg19" and "hg38" can be selected. Default is "rsID".}

\item{colors}{Optional. Vectors with 6 colors. "default" is `brewer.pal(6, "Set1")`.}
}
\value{
A Graph.
}
\description{
Return the network.
}
\examples{
plot_snp_network(snp="rs1059196")
}
